requirements.txt
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.7.0
dj-database-url==3.0.1
psycopg2-binary==2.9.10
Pillow==11.3.0
gunicorn==23.0.0
whitenoise==6.9.0
asgiref==3.9.1
sqlparse==0.5.3
tzdata==2025.2
django-environ==0.12.0
cloudinary==1.44.1
django-cloudinary-storage==0.3.0